requirements.txt
Django==5.0
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==3.5.0
djangorestframework-simplejwt==5.3.1
pytest==7.4.4
pytest_django==4.8.0
pytest_mock==3.12.0